Disable 32-bit-only moons
While Orbiter 2024 only fully supports 32-bit compilation, 64-bit is an alternate option that is offered and some people may wish to use it. To accommodate this, we will disable the celestial bodies which only exist as closed-source, 32-bit libraries, because they cause havoc with orbits and vessel trajectories when using the 64-bit version of Orbiter 2024. Disabling these only affects NASSP flights, and it will not impact optical sightings because none ever use the moons of the various other planets (nor do the optics have the capability to see them, in all honesty).